requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
lxml==5.3.0
numpy==2.1.2
openai==1.52.2
outcome==1.3.0.post0
pandas==2.2.3
pillow==10.4.0
pwinput==1.0.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
pdfplumber ==0.11.5
webdriver-auto-update==1.2.1
